+++ android/services/mongodb-remote-perftests/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.mongodb.stitch.android.services.mongodb.remote.perftests" />
+<manifest package="com.mongodb.stitch.android.services.mongodb.performance" />
