+++ core/services/mongodb-remote/src/test/java/com/mongodb/stitch/core/services/mongodb/remote/sync/internal/SyncUnitTestHarness.kt
@@ -716,6 +716,7 @@ class SyncUnitTestHarness : Closeable {
                 latestCtx?.localClient?.getDatabase(it)?.drop()
             }
         }
+
         latestCtx?.dataSynchronizer?.close()
     }
