+++ gradle.properties
@@ -6,7 +6,8 @@
 # http://www.gradle.org/docs/current/userguide/build_environment.html
 # Specifies the JVM arguments used for the daemon process.
 # The setting is particularly useful for tweaking memory settings.
-org.gradle.jvmargs=-Xmx8192m
+org.gradle.jvmargs=-Xmx1g
+JAVA_OPTS="-Xmx1gm -XX:MaxPermSize=1gm -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8"
 org.gradle.daemon=false
 # When configured, Gradle will run in incubating parallel mode.
 # This option should only be used with decoupled projects. More details, visit
